requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
colorama==0.4.1
config==0.4.2
Flask==1.0.3
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.1.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pytest==5.0.0
python-dotenv==0.10.3
six==1.12.0
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.1
